# docs/api-v1-openapi.yaml
openapi: 3.1.0
info:
  title: CloudOps API v1
  version: 1.0.0
  description: >-
    Local Owner commands and bounded MySQL-backed CloudOps projections.
    Numeric database identifiers, secrets, leases, checkpoints, and raw prompts
    are never public.
servers:
  - url: /
paths:
...
  /api/v1/alerts:
    get:
      operationId: listAlerts
      description: Independent Alert lifecycle page derived from immutable Signals.
      parameters:
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
        - { name: status, in: query, schema: { type: string, enum: [firing, resolved] } }
        - { name: severity, in: query, schema: { type: string, enum: [unknown, info, warning, critical] } }
        - { name: namespace, in: query, schema: { type: string, maxLength: 255 } }
        - { name: search, in: query, schema: { type: string, maxLength: 255 } }
      responses:
        "200":
          description: Bounded Alert page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AlertPage" }
        "400": { $ref: "#/components/responses/Problem" }
        "501": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/alerts/{id}:
    parameters:
      - { $ref: "#/components/parameters/PublicID" }
    get:
      operationId: getAlert
      responses:
        "200":
          description: Alert lifecycle, source Signal provenance and audit events
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AlertDetail" }
        "404": { $ref: "#/components/responses/Problem" }
        "501": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/alerts/{id}/acknowledgements:
    parameters:
      - { $ref: "#/components/parameters/PublicID" }
    post:
      operationId: acknowledgeAlert
      description: Local Owner records a recurrence-bound acknowledgement without silencing or resolving the Alert.
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKey" }
        - { $ref: "#/components/parameters/Origin" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AlertCommand" }
      responses:
        "200": { description: Updated Alert, content: { application/json: { schema: { $ref: "#/components/schemas/Alert" } } } }
        "400": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/alerts/{id}/silences:
    parameters:
      - { $ref: "#/components/parameters/PublicID" }
    post:
      operationId: createAlertSilence
      description: Local Owner creates an exact-matcher, provider-backed silence bounded to 5 minutes through 24 hours.
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKey" }
        - { $ref: "#/components/parameters/Origin" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AlertSilenceCommand" }
      responses:
        "201": { description: Provider-backed silence, content: { application/json: { schema: { $ref: "#/components/schemas/AlertSilence" } } } }
        "400": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/silences/{id}/expire:
    parameters:
      - { $ref: "#/components/parameters/PublicID" }
    post:
      operationId: expireAlertSilence
      description: Local Owner expires the provider silence while leaving Alert firing state unchanged.
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKey" }
        - { $ref: "#/components/parameters/Origin" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AlertExpireSilenceCommand" }
      responses:
        "200": { description: Expired silence, content: { application/json: { schema: { $ref: "#/components/schemas/AlertSilence" } } } }
        "400": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/alerts/{id}/investigations:
    parameters:
      - { $ref: "#/components/parameters/PublicID" }
    post:
      operationId: startAlertInvestigation
      description: Local Owner starts an Investigation through an explicitly linked active Incident.
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKey" }
        - { $ref: "#/components/parameters/Origin" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AlertCommand" }
      responses:
        "202": { description: Investigation accepted, content: { application/json: { schema: { $ref: "#/components/schemas/Alert" } } } }
        "400": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/alerts/{id}/incident-links:
    parameters:
      - { $ref: "#/components/parameters/PublicID" }
    post:
      operationId: linkAlertIncident
      description: Local Owner explicitly creates an Incident or attaches the Alert to an existing active Incident.
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKey" }
        - { $ref: "#/components/parameters/Origin" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AlertIncidentLinkCommand" }
      responses:
        "201": { description: Updated Alert relation, content: { application/json: { schema: { $ref: "#/components/schemas/Alert" } } } }
        "400": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
...
  /api/v1/incidents/{id}/investigations:
    parameters:
      - { $ref: "#/components/parameters/PublicID" }
    get:
      operationId: listInvestigations
      parameters: [{ $ref: "#/components/parameters/Cursor" }, { $ref: "#/components/parameters/Limit" }]
      responses:
        "200": { $ref: "#/components/responses/IncidentInvestigationPage" }
        "501": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
    post:
      operationId: startInvestigation
      description: Local Owner command with optimistic version and durable idempotency fencing.
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKey" }
        - { $ref: "#/components/parameters/Origin" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VersionedCommand" }
      responses:
        "202": { $ref: "#/components/responses/CommandAccepted" }
        "400": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "501": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
...
  /api/v1/devops:
    get:
      operationId: getDevOpsWorkspace
      description: MySQL projection of exact change, baseline, Plan, delivery, operation, provider, and Verification identities.
      parameters: [{ $ref: "#/components/parameters/Limit" }]
      responses:
        "200": { description: DevOps Workspace projection, content: { application/json: { schema: { $ref: "#/components/schemas/DevOpsWorkspace" } } } }
        "501": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
components:
  parameters:
    PublicID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid, minLength: 36, maxLength: 36 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string, maxLength: 256 }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
    Origin:
      name: Origin
      in: header
      required: true
      schema: { type: string, format: uri, maxLength: 2048 }
...
  schemas:
...
    IncidentInvestigation:
      type: object
      additionalProperties: false
      required: [id, cycle, status, version, objective, failure_summary, prompt_version, used_steps, max_steps, created_at, updated_at, context_link, migrated_legacy, migrated_legacy_context]
      properties:
        id: { type: string, format: uuid }
        cycle: { type: integer, minimum: 1 }
        status: { type: string, enum: [pending, running, completed, failed, cancelled] }
        version: { type: integer, minimum: 1 }
        objective: { type: string, maxLength: 2048 }
        outcome: { type: string, maxLength: 32 }
        failure_code: { type: string, maxLength: 128 }
        failure_summary: { type: string, maxLength: 2048 }
        model_provider: { type: string, maxLength: 64 }
        actual_model: { type: string, maxLength: 128 }
        prompt_version: { type: string, maxLength: 128 }
        used_steps: { type: integer, minimum: 0 }
        max_steps: { type: integer, minimum: 1 }
        started_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        context_link: { $ref: "#/components/schemas/ContextLink" }
        migrated_legacy: { type: boolean }
        migrated_legacy_context: { type: boolean }
...
    SHA256:
      type: string
      pattern: "^[0-9a-f]{64}$"
...
    AlertPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 100
          items: { $ref: "#/components/schemas/Alert" }
        next_cursor: { type: string }
    AlertDetail:
      type: object
      additionalProperties: false
      required: [alert, signals, events]
      properties:
        alert: { $ref: "#/components/schemas/Alert" }
        signals:
          type: array
          items: { $ref: "#/components/schemas/AlertSignal" }
        events:
          type: array
          items: { $ref: "#/components/schemas/AlertEvent" }
    Alert:
      type: object
      additionalProperties: false
      required: [id, status, severity, summary, category, source, fingerprint, correlation_key, cluster, environment, namespace, service_name, target_kind, target_name, first_seen_at, last_seen_at, starts_at, recurrence_count, signal_count, version, incident_links, investigations, context_link, migrated_legacy, migrated_legacy_context, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [firing, resolved] }
        severity: { type: string, enum: [unknown, info, warning, critical] }
        summary: { type: string, maxLength: 2048 }
        category: { type: string, maxLength: 255 }
        source: { type: string, maxLength: 64 }
        fingerprint: { type: string, maxLength: 128 }
        correlation_key: { $ref: "#/components/schemas/SHA256" }
        cluster: { type: string, maxLength: 128 }
        environment: { type: string, maxLength: 64 }
        namespace: { type: string, maxLength: 255 }
        service_name: { type: string, maxLength: 255 }
        target_kind: { type: string, maxLength: 64 }
        target_name: { type: string, maxLength: 255 }
        first_seen_at: { type: string, format: date-time }
        last_seen_at: { type: string, format: date-time }
        starts_at: { type: string, format: date-time }
        resolved_at: { type: [string, "null"], format: date-time }
        recurrence_count: { type: integer, minimum: 1 }
        signal_count: { type: integer, minimum: 1 }
        version: { type: integer, minimum: 1 }
        acknowledgement: { oneOf: [{ $ref: "#/components/schemas/AlertAcknowledgement" }, { type: "null" }] }
        silence: { oneOf: [{ $ref: "#/components/schemas/AlertSilence" }, { type: "null" }] }
        incident_links:
          type: array
          items: { $ref: "#/components/schemas/AlertIncidentLink" }
        investigations:
          type: array
          items: { $ref: "#/components/schemas/AlertInvestigation" }
        context_link: { $ref: "#/components/schemas/AlertContextLink" }
        migrated_legacy: { type: boolean }
        migrated_legacy_context: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    AlertAcknowledgement:
      type: object
      additionalProperties: false
      required: [id, recurrence_no, alert_version, actor, reason, created_at]
      properties:
        id: { type: string, format: uuid }
        recurrence_no: { type: integer, minimum: 1 }
        alert_version: { type: integer, minimum: 1 }
        actor:
          type: object
          additionalProperties: false
          required: [provider, login, role]
          properties:
            provider: { type: string, const: local }
            login: { type: string, const: owner }
            role: { type: string, const: owner }
        reason: { type: string, minLength: 1, maxLength: 1024 }
        created_at: { type: string, format: date-time }
    AlertMatcher:
      type: object
      additionalProperties: false
      required: [name, value, is_regex, is_equal]
      properties:
        name: { type: string, maxLength: 128 }
        value: { type: string, maxLength: 1024 }
        is_regex: { type: boolean, const: false }
        is_equal: { type: boolean, const: true }
    AlertSilence:
      type: object
      additionalProperties: false
      required: [id, status, matchers, reason, configuration_revision_id, starts_at, ends_at, created_at]
      properties:
        id: { type: string, format: uuid }
        provider_silence_id: { type: string, maxLength: 128 }
        status: { type: string, enum: [pending, active, expired, failed] }
        matchers:
          type: array
          minItems: 1
          maxItems: 8
          items: { $ref: "#/components/schemas/AlertMatcher" }
        reason: { type: string, minLength: 1, maxLength: 1024 }
        configuration_revision_id: { type: string, format: uuid }
        starts_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time }
        expired_at: { type: [string, "null"], format: date-time }
        provider_error_code: { type: string, maxLength: 64 }
        created_at: { type: string, format: date-time }
    AlertIncidentLink:
      type: object
      additionalProperties: false
      required: [id, incident_id, incident_cycle, incident_status, provenance, created_at]
      properties:
        id: { type: string, format: uuid }
        incident_id: { type: string, format: uuid }
        incident_cycle: { type: integer, minimum: 1 }
        incident_status: { type: string }
        provenance: { type: string, enum: [owner_created, owner_attached, escalation_policy, legacy_automatic_ingress] }
        configuration_revision_id: { type: string, format: uuid }
        escalation_policy_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
    AlertInvestigation:
      type: object
      additionalProperties: false
      required: [id, incident_id, status, created_at]
      properties:
        id: { type: string, format: uuid }
        incident_id: { type: string, format: uuid }
        status: { type: string, enum: [pending, running, completed, failed, cancelled] }
        created_at: { type: string, format: date-time }
    AlertContextLink:
      type: object
      additionalProperties: false
      required: [workspace, path, query, operational_scope_id, external]
      properties:
        workspace: { type: string, const: alerts }
        path: { type: string, pattern: "^/alerts/[0-9a-f-]{36}$" }
        query: { type: object, additionalProperties: { type: string } }
        operational_scope_id: { type: string }
        external: { type: boolean, const: false }
    AlertSignal:
      type: object
      additionalProperties: true
      required: [id, source_event_id, alert_instance_key, status, severity, summary, labels, annotations, starts_at, occurred_at, received_at, provenance]
      properties:
        id: { type: string, format: uuid }
        source_event_id: { $ref: "#/components/schemas/SHA256" }
        alert_instance_key: { $ref: "#/components/schemas/SHA256" }
        status: { type: string, enum: [firing, resolved] }
        severity: { type: string }
        summary: { type: string }
        labels: { type: object }
        annotations: { type: object }
        starts_at: { type: string, format: date-time }
        ends_at: { type: [string, "null"], format: date-time }
        occurred_at: { type: string, format: date-time }
        received_at: { type: string, format: date-time }
        provenance: { type: string, enum: [signal_normalization, legacy_automatic_ingress] }
    AlertEvent:
      type: object
      additionalProperties: false
      required: [id, type, actor_type, actor_id, summary, metadata, occurred_at]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, maxLength: 64 }
        actor_type: { type: string, maxLength: 32 }
        actor_id: { type: string, maxLength: 128 }
        summary: { type: string, maxLength: 2048 }
        metadata: { type: object }
        occurred_at: { type: string, format: date-time }
    AlertCommand:
      type: object
      additionalProperties: false
      required: [expected_version, reason]
      properties:
        expected_version: { type: integer, minimum: 1 }
        reason: { type: string, minLength: 1, maxLength: 1024 }
    AlertSilenceCommand:
      type: object
      additionalProperties: false
      required: [expected_version, duration_seconds, reason]
      properties:
        expected_version: { type: integer, minimum: 1 }
        duration_seconds: { type: integer, minimum: 300, maximum: 86400 }
        reason: { type: string, minLength: 1, maxLength: 1024 }
    AlertExpireSilenceCommand:
      type: object
      additionalProperties: false
      required: [expected_version]
      properties:
        expected_version: { type: integer, minimum: 1 }
    AlertIncidentLinkCommand:
      type: object
      additionalProperties: false
      required: [expected_version, create]
      properties:
        expected_version: { type: integer, minimum: 1 }
        create: { type: boolean }
        incident_id: { type: string, format: uuid }
    VersionedCommand:
      type: object
      additionalProperties: false
      required: [expected_version]
      properties:
        expected_version: { type: integer, minimum: 1 }
        reason: { type: string, maxLength: 2048 }
...
    DevOpsWorkspace:
      type: object
      description: Durable DevOps projection. Operation Plan and Action Card objects retain their Agent authority schema.
      required: [operation_plans, action_cards, executions, change_freezes, change_candidates, deployment_baselines, deliveries, providers, collected_at]
      properties:
        operation_plans: { type: array, items: { $ref: "#/components/schemas/OperationPlan" } }
        action_cards: { type: array, items: { type: object } }
        executions: { type: array, items: { $ref: "#/components/schemas/OperationExecution" } }
        change_freezes: { type: array, items: { $ref: "#/components/schemas/ChangeFreezeState" } }
        change_candidates: { type: array, items: { type: object } }
        deployment_baselines: { type: array, items: { type: object } }
        deliveries: { type: array, items: { type: object } }
        providers: { type: array, items: { type: object } }
        collected_at: { type: string, format: date-time }
...
    CommandResult:
      type: object
      additionalProperties: false
      required: [id, command, status]
      properties:
        id: { type: string, format: uuid }
        command: { type: string }
        status: { type: string }
        version: { type: integer, minimum: 1 }
        cycle: { type: integer, minimum: 1 }
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance, code, request_id, trace_id, next_steps]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
        request_id: { type: string }
        trace_id: { type: string }
        next_steps:
          type: array
          minItems: 1
          items: { type: string }
  responses:
...
    IncidentInvestigationPage:
      description: Bounded current-cycle Agent Investigation page
      content:
        application/json:
          schema:
            type: object
            required: [items]
            properties:
              items:
                type: array
                maxItems: 100
                items: { $ref: "#/components/schemas/IncidentInvestigation" }
              next_cursor: { type: string, format: uuid }
...
    CommandAccepted:
      description: Command persisted and queued
      headers:
        X-Request-ID: { schema: { type: string } }
        X-Trace-ID: { schema: { type: string } }
        Idempotent-Replay: { schema: { type: string, enum: ["true"] } }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/CommandResult" }
    Problem:
      description: Stable problem response
      headers:
        X-Request-ID: { schema: { type: string } }
        X-Trace-ID: { schema: { type: string } }
        Idempotent-Replay: { schema: { type: string, enum: ["true"] } }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }